+++ client/src/css/insight.css
