+++ client/src/utils/mutations.js
@@ -12,13 +12,13 @@ export const ADD_USER = gql`
     }
 `;
 
-export const LOGIN = gql`
+export const LOGIN_USER = gql`
     mutation login($email: String!, $password: String!) {
         login(email: $email, password: $password) {
             token
-            profile {
-                _id
+            user {
                 email
+                password
             }
         }
     }
