+++ server/seeders/userSeeds.json
@@ -1,19 +1,19 @@
 [
   {
-    "first_name": "Scott",
-    "last_name": "Erwin",
+    "firstName": "Scott",
+    "lastName": "Erwin",
     "email": "scotterwin@gmail.com",
     "password": "yousuck"
   },
   {
-    "first_name": "Richard",
-    "last_name": "Crippen",
+    "firstName": "Richard",
+    "lastName": "Crippen",
     "email": "richard.crippen@gmail.com",
     "password": "abc123"
   },
   {
-    "first_name": "Zack",
-    "last_name": "Eggert",
+    "firstName": "Zack",
+    "lastName": "Eggert",
     "email": "zreggert@gmail.com",
     "password": "kittycat"
   }
